+++ ProyectoCampeonato/src/main/java/org/example/proyectocampeonato/repository/CompetidorRepository.java
@@ -3,5 +3,5 @@ package org.example.proyectocampeonato.repository;
 import org.example.proyectocampeonato.modelo.Competidor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompetidorRepository extends JpaRepository<Competidor, Integer> {
+public interface CompetidorRepository extends JpaRepository<Competidor, Long> {
 }
